fixed rating bar step -> now user cannot enter decimal scores

+++ app/src/main/java/com/rayofdoom/shows_leo/ShowDetailsActivity.kt
@@ -113,7 +113,7 @@ class ShowDetailsActivity : AppCompatActivity() {
                             ).toString(),
                         dialogBinding.reviewInput.text.toString(),
                         R.drawable.ic_profile_placeholder,
-                        dialogBinding.rating.rating.toDouble()
+                        dialogBinding.rating.rating.toInt()
                     )
                 )
                 reviewsAdapter?.addItem(reviews)
